Added the service integration and tests

+++ src/main/java/uk/gav/pun/repository/PunRepository.java
@@ -5,6 +5,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import uk.gav.pun.entity.Pun;
 
 
-public interface PunRepository  extends JpaRepository<Pun, Integer> {
+public interface PunRepository  extends JpaRepository<Pun, Long> {
 
 }
